minor bugs in parser fixed

+++ Rectangle.js
@@ -5,7 +5,7 @@
  */
 function Rectangle(scene, coords) {
 	CGFobject.call(this,scene);
-	
+
 	this.coords = coords;
 
 	this.initBuffers();
@@ -16,21 +16,21 @@ Rectangle.prototype = Object.create(CGFobject.prototype);
 Rectangle.prototype.constructor=Rectangle;
 
 Rectangle.prototype.initBuffers = function () {
-	
+
 	this.vertices = [];
-    this.vertices.push(parseInt(this.coords[0]), parseInt(this.coords[1]), 0);
-    this.vertices.push(parseInt(this.coords[2]), parseInt(this.coords[1]), 0);
-	this.vertices.push(parseInt(this.coords[2]), parseInt(this.coords[3]), 0);
-	this.vertices.push(parseInt(this.coords[0]), parseInt(this.coords[3]), 0);
-	
-	
+    this.vertices.push(parseFloat(this.coords[0]), parseFloat(this.coords[1]), 0);
+    this.vertices.push(parseFloat(this.coords[2]), parseFloat(this.coords[1]), 0);
+	this.vertices.push(parseFloat(this.coords[2]), parseFloat(this.coords[3]), 0);
+	this.vertices.push(parseFloat(this.coords[0]), parseFloat(this.coords[3]), 0);
+
+
 	this.indices = [
 		0, 1, 2,
 		2, 3, 0,
 		2, 1, 0,
 		0, 3, 2,
 	];
-		
+
 
 	this.normals = [
 		0, 0, 1,
@@ -43,7 +43,7 @@ Rectangle.prototype.initBuffers = function () {
 	this.maxS = 1;
 	this.minT = 0;
 	this.maxT = 1;
-	
+
 	this.texCoords = [
 		0, 0,
 		1, 0,
@@ -56,7 +56,7 @@ Rectangle.prototype.initBuffers = function () {
 };
 
 Rectangle.prototype.setTextCoords = function(s,t){
-	
+
 	this.texCoords = [
 		0, 0,
 		(this.coords[2]-this.coords[0])/s, 0,
@@ -65,5 +65,5 @@ Rectangle.prototype.setTextCoords = function(s,t){
 	];
 
 	this.updateTexCoordsGLBuffers();
-    
+
 };
